feat: clean controllers and services
