+++ server/src/core/auth/auth.module.ts
@@ -3,12 +3,15 @@ import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { JwtAuthStrategy } from './strategies/jwt.auth.strategy';
 import { AuthService } from './services/auth.service';
-import { GoogleAuthController } from './google.auth.controller';
+import { GoogleAuthController } from './controllers/google-auth.controller';
 import { GoogleStrategy } from './strategies/google.auth.strategy';
-import { TokenController } from './token.controller';
+import { TokenController } from './controllers/token.controller';
 import { PrismaService } from 'src/database/prisma.service';
 import { UserModule } from '../user/user.module';
-import { AuthController } from './auth.controller';
+import { AuthController } from './controllers/auth.controller';
+import { PasswordAuthController } from './controllers/password-auth.controller';
+
+import { TokenService } from './services/token.service';
 
 const jwtModule = JwtModule.registerAsync({
   useFactory: async (configService: ConfigService) => {
@@ -25,8 +28,19 @@ const jwtModule = JwtModule.registerAsync({
 
 @Module({
   imports: [jwtModule, ConfigModule.forRoot({}), UserModule],
-  controllers: [GoogleAuthController, TokenController, AuthController],
-  providers: [AuthService, JwtAuthStrategy, GoogleStrategy, PrismaService],
+  controllers: [
+    GoogleAuthController,
+    PasswordAuthController,
+    TokenController,
+    AuthController,
+  ],
+  providers: [
+    AuthService,
+    TokenService,
+    JwtAuthStrategy,
+    GoogleStrategy,
+    PrismaService,
+  ],
   exports: [jwtModule],
 })
 export class AuthModule {}
